Making sure we get the indexes of thr right element
In cases like `var [x] = 1;` we were generating `var x = undefined;` but this
is not acurate according to firefox's implementation. It should get the
`0` property and assign it to x in this case.
A test case was created to make sure that the final value of `x` in this case
is really the property `0` of `1`.
Also fixed a bug on the way we generate temporary variables.